+++ docs/changes.rst
@@ -6,6 +6,7 @@ Weblate 5.11
 .. rubric:: New features
 
 * Added :http:get:`/api/units/(int:id)/translations/` to retrieve a list of all target translation units for the given source translation unit.
+* Added :http:delete:`/api/groups/(int:id)/roles/(int:role_id)` to delete a role from a group.
 
 .. rubric:: Improvements